bookmark_manager_service/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<3.15
gunicorn>=21,<22
mysqlclient>=2.1,<2.3
python-dotenv>=1.0
selenium>=4.15,<5.0
webdriver-manager>=4.0,<5.0
requests>=2.31,<3.0
pillow>=10.0,<11.0
django-cors-headers>=4.0,<5.0
celery>=5.3,<6.0
redis>=5.0,<6.0
# Add other bookmark-specific dependencies here